Refactored the data structure a bit
