+++ src/buffer/channel.rs
@@ -503,8 +503,8 @@ mod nick_list {
             let content = selectable_text(user.display(nicklist_config.show_access_levels))
                 .style(|theme| theme::selectable_text::nicklist_nickname(theme, config, user))
                 .horizontal_alignment(match nicklist_config.alignment {
-                    config::channel::Alignment::Left => alignment::Horizontal::Left,
-                    config::channel::Alignment::Right => alignment::Horizontal::Right,
+                    config::buffer::channel::Alignment::Left => alignment::Horizontal::Left,
+                    config::buffer::channel::Alignment::Right => alignment::Horizontal::Right,
                 })
                 .width(Length::Fixed(width));
